refactor: NF-T4 Float 类型统一消除 DB↔FSRS 精度损失
MemoRecordEntity / ReviewLogEntity 的 stability/difficulty 由 Double 改为 Float,
与 FSRS FlashCard 类型完全一致。MemoRecordMapper 转换由 Float↔Double 双向变恒等,
彻底消除 NF-T4 报告的精度损失(原 1.5f → 1.5000000476837158 → 1.5f 的往返)。

变更范围:
- core:database: MemoRecordEntity / ReviewLogEntity Double→Float(REAL affinity 不变,schema hash 不变,无需 v5→v6 迁移)
- core:data: MemoRecordMapper 移除 .toFloat()/.toDouble() 转换;SchedulingRepository / SeedDataLoader / GraphRepositoryImpl 同步更新
- feature:aiassistant: AiAssistantViewModelTest 8 处 ReviewLogEntity 字面量 10.0→10f / 5.0→5f / 1.0→1f / 8.0→8f
- 新增 MemoRecordMapperTest: 11 个用例(round-trip 一致性 + 边界值 + 默认值 + state 枚举映射 + NF-T4 无精度损失验证)

验证: assembleDebug SUCCESSFUL + 231 tests 0 failures(220 基线 + 11 新)

+++ core/data/src/main/java/com/wenyan/app/core/data/mapper/MemoRecordMapper.kt
@@ -60,8 +60,8 @@ object MemoRecordMapper {
 
         return FlashCard(
             dueDate = dueDate,
-            stability = memoRecord.stability.toFloat(),
-            difficulty = memoRecord.difficulty.toFloat(),
+            stability = memoRecord.stability,
+            difficulty = memoRecord.difficulty,
             interval = if (memoRecord.lastReviewAt > 0) {
                 ((memoRecord.nextReviewAt - memoRecord.lastReviewAt) / DAY_MS).toInt()
             } else 0,
@@ -103,8 +103,8 @@ object MemoRecordMapper {
         return MemoRecordEntity(
             pointId = pointId,
             state = flashCard.state.name,
-            stability = flashCard.stability.toDouble(),
-            difficulty = flashCard.difficulty.toDouble(),
+            stability = flashCard.stability,
+            difficulty = flashCard.difficulty,
             lastReviewAt = lastReviewAt,
             nextReviewAt = nextReviewAt,
             reviewCount = flashCard.reviewCount,
